Refactor focus ring implementation for dark themes and add tests
