+++ themes/dark/theme.html
@@ -27,7 +27,7 @@ limitations under the License.
 
       /* Brighter keyboard focus ring so it stays visible on the dark backgrounds (WCAG 2.4.13). */
       --nuxeo-focus-ring-color: #4d9fff;
-      /* Full outline consumed by the global focus ring (themes/focus-ring.js). Defining it here makes
+      /* Full outline consumed by the global focus ring (themes/dark-theme-focus-ring.js). Defining it here makes
          keyboard focus visible app-wide on dark backgrounds; other themes leave it unset and keep the
          native user-agent ring. */
       --nuxeo-focus-outline: 2px solid var(--nuxeo-focus-ring-color);
