+++ themes/hyland-dark/theme.html
@@ -28,7 +28,7 @@ limitations under the License.
       --nuxeo-primary-color: #c3c0ff;
       --nuxeo-secondary-color: #948df6;
 
-      /* Full outline consumed by the global focus ring (themes/focus-ring.js). Defining it here makes
+      /* Full outline consumed by the global focus ring (themes/dark-theme-focus-ring.js). Defining it here makes
          keyboard focus visible app-wide on dark backgrounds; other themes leave it unset and keep the
          native user-agent ring. */
       --nuxeo-focus-outline: 2px solid var(--nuxeo-primary-color);
